+++ apps/cron/src/index.ts
@@ -27,6 +27,10 @@ await sendMessage(
   `## Portfolio Daily Cron: Run Completed
 App Version: ${APP_VERSION} ${environment.DRY_RUN ? "(Dry Run)" : ""}
 ${await getSummary()}${
+    logger.hasEstimation
+      ? "\n📐 Estimations were made on some asset/currency price."
+      : ""
+  }${
     logger.hasWarning ? "\n⚠️ Warnings were found during the run." : ""
   }${logger.hasError ? "\n❗ Errors were found during the run." : ""}`,
   logger.getMessages().join("\n"),