+++ apps/cron/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 0.1.0 is not noted here
 
+## [0.6.0] - 2025-12-12
+
+- fix: currency `updatedAt` column not updated
+- add notice for estimation
+
 ## [0.5.0] - 2025-12-11
 
 - using usdt/thb to estimate currency exchange rate